Revert "Date enhancement"
