--- Sources/Swift/Components/Date/TimeZone+Extensions.swift
@@ -1,11 +1,0 @@
-//
-// Xcore
-// Copyright © 2014 Xcore
-// MIT license, see LICENSE file for details
-//
-
-import Foundation
-
-extension TimeZone {
-    public static let gmt = TimeZone(identifier: "GMT")!
-}
